Use failure_notifier instead of Formatter
Initially, I thought the only way to detect a test failure in RSpec was to use Formatter. This approach was counter-intuitive, but I had no choice because I didn't think there was any other way to do it.

However, I realized that RSpec could use RSpec::Support.failure_notifier, so I switched to this method. This should make the code easier to understand.